requirements.txt
openai-whisper
fastapi
python-multipart
uvicorn
ffmpeg-python
celery